+++ selfdrive/test/process_replay/__init__.py
@@ -1,0 +1,1 @@
+from selfdrive.test.process_replay.process_replay import CONFIGS, get_process_config, replay_process, replay_process_with_name  # noqa: F401
