+++ selfdrive/test/process_replay/compare_logs.py
@@ -1,5 +1,4 @@
 #!/usr/bin/env python3
-import bz2
 import sys
 import math
 import capnp
@@ -10,16 +9,6 @@ from collections import Counter
 from tools.lib.logreader import LogReader
 
 EPSILON = sys.float_info.epsilon
-
-
-def save_log(dest, log_msgs, compress=True):
-  dat = b"".join(msg.as_builder().to_bytes() for msg in log_msgs)
-
-  if compress:
-    dat = bz2.compress(dat)
-
-  with open(dest, "wb") as f:
-    f.write(dat)
 
 
 def remove_ignored_fields(msg, ignore):
