+++ tools/lib/helpers.py
@@ -1,3 +1,4 @@
+import bz2
 import datetime
 
 TIME_FMT = "%Y-%m-%d--%H-%M-%S"
@@ -13,8 +14,19 @@ class RE:
   EXPLORER_FILE = r'^(?P<segment_name>{})--(?P<file_name>[a-z]+\.[a-z0-9]+)$'.format(SEGMENT_NAME)
   OP_SEGMENT_DIR = r'^(?P<segment_name>{})$'.format(SEGMENT_NAME)
 
+
 def timestamp_to_datetime(t: str) -> datetime.datetime:
   """
     Convert an openpilot route timestamp to a python datetime
   """
   return datetime.datetime.strptime(t, TIME_FMT)
+
+
+def save_log(dest, log_msgs, compress=True):
+  dat = b"".join(msg.as_builder().to_bytes() for msg in log_msgs)
+
+  if compress:
+    dat = bz2.compress(dat)
+
+  with open(dest, "wb") as f:
+    f.write(dat)
